Add FSpDelete; Delete/HDelete should delete files or folders.

+++ toolbox/os.cpp
@@ -608,6 +608,8 @@ namespace OS
 			_ioDirID = 48,
 		};
 
+		struct stat st;
+
 		bool htrap = trap & 0x0200;
 		const char *func = htrap ? "HDelete" : "Delete";
 
@@ -635,7 +637,17 @@ namespace OS
 
 		Log("     %s(%s)\n", func, sname.c_str());
 
-		int ok = ::unlink(sname.c_str());
+		int ok;
+
+		ok = ::stat(sname.c_str(), &st);
+		if (ok == 0)
+		{
+			if (S_ISDIR(st.st_mode))
+				ok = ::rmdir(sname.c_str());
+			else
+				ok = ::unlink(sname.c_str());
+		}
+
 		if (ok < 0)
 			d0 = macos_error_from_errno();
 		else
